+++ CHANGELOG.rst
@@ -13,5 +13,14 @@ Version 0.1.0 (2019-09-08)
 * Added: initial release
 
 
+Version 0.1.1 (2019-09-23)
+--------------------------
+
+* Added: advanced usage example
+* Added: ``AudioBuffer.from_file()`` to read from an audio file
+* Added: ``AudioBuffer.to_file()`` to save buffer to a an audio file
+* Changed: ``AudioBuffer.FromArray()`` to ``AudioBuffer.from_array()``
+
+
 .. _Keep a Changelog: https://keepachangelog.com/en/1.0.0/
 .. _Semantic Versioning: https://semver.org/spec/v2.0.0.html